+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog for the Mapbox Navigation SDK Core Framework for Android
 
+## Navigation SDK Core Framework 3.4.0-rc.1 - 17 September, 2024
+#### Features
+
+
+#### Bug fixes and improvements
+- Fixed `CarSearchLocationProvider` produces _NullPointerException_ when using Mapbox Search SDK.  [#6702](https://github.com/mapbox/mapbox-navigation-android/pull/6702)
+
+#### Known issues :warning:
+
+
+#### Other changes
+
+
+### Mapbox dependencies
+This release depends on, and has been tested with, the following Mapbox dependencies:
+- Mapbox Maps SDK `v11.7.0-rc.1` ([release notes](https://github.com/mapbox/mapbox-maps-android/releases/tag/v11.7.0-rc.1))
+- Mapbox Navigation Native `v319.0.0`
+- Mapbox Core Common `v24.7.0-rc.2`
+- Mapbox Java `v7.2.0` ([release notes](https://github.com/mapbox/mapbox-java/releases/tag/v7.2.0))
+
+
 ## Navigation SDK Core Framework 3.4.0-beta.1 - 05 September, 2024
 #### Features
 - Signature of experimental `EtcGateApi#updateEtcGateInfo` function has been changed, now it accepts `EtcGateApi.EtcGateInfo` as a function parameter. [#6508](https://github.com/mapbox/mapbox-navigation-android/pull/6508)
