--- CODEOWNERS
@@ -1,1 +1,0 @@
-* @mapbox/navigation-android
