+++ libnavigation-android/build.gradle
@@ -36,4 +36,3 @@ dependencies {
 apply from: "../gradle/track-public-apis.gradle"
 apply from: "../gradle/jacoco.gradle"
 apply from: "../gradle/publish.gradle"
-apply from: "../../../../scripts_shared/find-all-common-sdk-versions.gradle"
