+++ libnavigator/src/androidTest/java/com/mapbox/navigation/navigator/internal/NavigatorTest.kt
@@ -123,5 +123,6 @@ class NavigatorTest {
         null,
         null,
         null,
+        null,
     )
 }
