+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/internal/route/callout/model/DurationDifferenceType.kt
@@ -1,0 +1,8 @@
+package com.mapbox.navigation.ui.maps.internal.route.callout.model
+
+import androidx.annotation.RestrictTo
+
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
+enum class DurationDifferenceType {
+    Faster, Slower, Same
+}
