--- libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/route/callout/model/DurationDifferenceType.kt
@@ -1,5 +1,0 @@
-package com.mapbox.navigation.ui.maps.route.callout.model
-
-internal enum class DurationDifferenceType {
-    Faster, Slower, Same
-}
